+++ naga/src/common/mod.rs
@@ -1,0 +1,3 @@
+//! Code common to the front and backends for specific languages.
+
+pub mod wgsl;
